requirements.txt
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
ipython==8.18.0
psycopg2==2.9.9